requirements-test.txt
# Test dependencies
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0

# Load testing
locust>=2.15.0

# API testing
requests>=2.28.0
httpx>=0.24.0

# Security testing
bandit>=1.7.0
safety>=2.3.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Coverage reporting
coverage>=7.0.0
